Add check on NA values
I had NA data in my dataset, which created this error: Error in if (any(decOccured)) { : missing value where TRUE/FALSE needed
It was a bit tedious before I founded out it was due to the NA values. 
I suggest for other people with forgotten NA values another check if any values with NA values are present in data.
I added it and kept the style according to the other warnings.